requirements.txt
# DataPilot v3 — Full dependency list
# pip install -r requirements.txt

# ── Web framework ────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# ── Data science ─────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
scipy>=1.12.0
pyarrow>=15.0.0        # for .parquet dataset storage

# ── Auth ─────────────────────────────────────────────────
PyJWT>=2.8.0
bcrypt>=4.1.0

# ── Payments ─────────────────────────────────────────────
stripe>=8.0.0

# ── Database ─────────────────────────────────────────────
asyncpg>=0.29.0        # PostgreSQL async driver
# boto3>=1.34.0        # S3 file storage (uncomment for production)

# ── HTTP client (Claude API fallback) ────────────────────
httpx>=0.27.0
